fix technic
fixed technic with the LaunchClassLoader patch again. I have no idea why but calling it during preinit usually makes technic use 398MB wrap around point for tekkit compared to without a 530MB wrap around point for the ram and it was tested consistently. The old version said it was dummy maps all the way through and the cl's were the same instances so it shouldn't happen they are on some weird shiz