remove broken code
sadly not possible to get around windows bug as it's locked out and only appears to work if the cmd is the parent process of change_niceness.exe and if and only if conhost.exe is the parent process of cmd. It's too much of a mess and this hack could no longer work in the near future thus we removed it

we enforce javaagent always for 1.5x now for high process priority it's not required if it's disabled.