allow legacy patch loaders
allow patching of RelaunchClassLoader and Technic's MinecraftClassLoader (if not bugged). when strict mode is off it means the user turned it off of auto and you are 1.5x or below so allow the agressive patching to happen as theortically any replacement of the 1.5x classloader will have the same issue as them.